Create unsaved team object during team creation
This is going to be necessary to ensure contact uniqueness
ie. that if person A is on teams X and Y, person A only exists once in storage

Longer explanation:
If we don't have a team object, then we just know that we're creating person P.
If person P is already on another team, then we need to make sure that we only add new memberships to the new team.
Rather than query to find such memberships, it's simpler to just create a team and save when ready.