Make address book access checks asynchronous
We need to handle this manually because in iOS 8, accessing
the people picker no longer automatically requests AB access.
Instead, if the app does not have access, copies of contacts
will be quietly returned (who have invalid IDs, preventing
contact-syncing).

Now we ensure address book access, but this access check is async, so
we have to pass in what we want to do next as callbacks.

Resolves issue #13